condaenv.gvixhd5a.requirements.txt
absl-py==0.12.0
alembic==1.6.0
apispec==4.4.0
apispec-webframeworks==0.5.2
appdirs==1.4.4
apscheduler==3.7.0
astor==0.8.1
astunparse==1.6.3
audioread==2.1.9
azure-cognitiveservices-speech==1.17.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bidict==0.21.2
boto3==1.17.64
botocore==1.20.64
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cython==0.29.23
decorator==4.4.2
deprecated==1.2.12
dnspython==1.16.0
editdistance==0.5.3
email-validator==1.1.0
eventlet==0.25.2
flask==1.1.2
flask-bcrypt==0.7.1
flask-cors==3.0.8
flask-httpauth==4.0.0
flask-jsglue==0.3.1
flask-jwt-extended==4.1.0
flask-login==0.5.0
flask-migrate==3.1.0
flask-socketio==5.0.0
flask-sqlalchemy==2.4.1
flask-swagger-ui==3.36.0
flask-wtf==0.14.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==20.5.2
gevent-websocket==0.10.1
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.34.1
h5py==3.1.0
idna==2.9
imageio==2.9.0
importlib-metadata==4.0.1
imutils==0.5.3
itsdangerous==1.1.0
jaconv==0.3
jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.0.0
keras-applications==1.0.8
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.7.0
llvmlite==0.31.0
mako==1.1.4
markdown==3.3.4
markupsafe==1.1.1
marshmallow==3.11.1
matplotlib==3.3.4
mecab-python3==1.0.4
monotonic==1.5
networkx==2.5.1
numba==0.48.0
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==3.4.2.16
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.9
pandas==0.23.3
passlib==1.7.4
pdf2image==1.13.1
pillow==7.2.0
pip
pooch==1.3.0
protobuf==3.17.1
py3dbp==1.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyjwt==2.1.0
pykakasi==2.2.1
pymysql==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
pywavelets==1.1.1
pyyaml==6.0
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.1
shapely==1.7.0
six==1.15.0
soundfile==0.10.3.post1
soupsieve==2.2.1
sqlalchemy==1.3.16
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
tinydb==4.4.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.47.0
typing-extensions==3.7.4.3
tzlocal==2.1
unidic-lite==1.0.8
urllib3==1.25.9
werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
wtforms==2.3.1
xlwt==1.3.0
zipp==3.4.1
zope-event==4.4
zope-interface==5.1.0